# docs/swagger.yaml
basePath: /
definitions:
...
  model.CaseHistoryInsert:
    properties:
      caseId:
        type: string
      fullMsg:
        type: string
      jsonData:
        type: string
      type:
        type: string
      username:
        type: string
    type: object
  model.CaseHistoryUpdate:
    properties:
      fullMsg:
        type: string
      jsonData:
        type: string
      type:
        description: |-
          CaseID   string          `json:"caseId" db:"caseId"`
          Username string          `json:"username" db:"username"`
        type: string
    type: object
  model.CaseInsert:
    properties:
      arrivedDate:
        type: string
      attachments:
        items:
          $ref: '#/definitions/model.TixCaseAttachmentInput'
        type: array
      caseDetail:
        type: string
      caseDuration:
        type: integer
      caseId:
        type: string
      caseLat:
        type: string
      caseLon:
        type: string
      caseSTypeId:
        type: string
      caseTypeId:
        type: string
      caseVersion:
        type: string
      caselocAddr:
        type: string
      caselocAddrDecs:
        type: string
      closedDate:
        type: string
      commandedDate:
        type: string
      countryId:
        type: string
      createdDate:
        type: string
      deviceId:
        type: string
      distId:
        type: string
      extReceive:
        type: string
      formData:
        $ref: '#/definitions/model.FormAnswerRequest'
      nodeId:
        type: string
      phoneNo:
        type: string
      phoneNoHide:
        type: boolean
      priority:
        type: integer
      provId:
        type: string
      receivedDate:
        type: string
      referCaseId:
        type: string
      resDetail:
        type: string
      resId:
        type: string
      scheduleDate:
        type: string
      scheduleFlag:
        type: boolean
      source:
        type: string
      startedDate:
        type: string
      statusId:
        type: string
      userarrive:
        type: string
      userclose:
        type: string
      usercommand:
        type: string
      usercreate:
        type: string
      userreceive:
        type: string
      versions:
        type: string
      wfId:
        type: string
    required:
    - caseVersion
    type: object
...
  model.CaseUpdate:
    properties:
      arrivedDate:
        type: string
      caseDetail:
        type: string
      caseDuration:
        type: integer
      caseId:
        type: string
      caseLat:
        type: string
      caseLon:
        type: string
      caseSTypeId:
        type: string
      caseTypeId:
        type: string
      caseVersion:
        type: string
      caselocAddr:
        type: string
      caselocAddrDecs:
        type: string
      closedDate:
        type: string
      commandedDate:
        type: string
      countryId:
        type: string
      createdDate:
        type: string
      deviceId:
        type: string
      distId:
        type: string
      extReceive:
        type: string
      phoneNo:
        type: string
      phoneNoHide:
        type: boolean
      priority:
        type: integer
      provId:
        type: string
      receivedDate:
        type: string
      referCaseId:
        type: string
      resDetail:
        type: string
      resId:
        type: string
      scheduleDate:
        type: string
      scheduleFlag:
        type: boolean
      source:
        type: string
      startedDate:
        type: string
      statusId:
        type: string
      userarrive:
        type: string
      userclose:
        type: string
      usercommand:
        type: string
      usercreate:
        type: string
      userreceive:
        type: string
      wfId:
        type: string
    type: object
...
  model.FormAnswerRequest:
    properties:
      formColSpan:
        type: integer
      formFieldJson:
        items:
          additionalProperties: true
          type: object
        type: array
      formId:
        type: string
      formName:
        type: string
      nextNodeId:
        type: string
      versions:
        type: string
      wfId:
        type: string
    type: object
...
  model.Response:
    properties:
      data: {}
      desc:
        type: string
      msg:
        type: string
      status:
        type: string
    type: object
...
  model.TixCaseAttachmentInput:
    properties:
      attId:
        type: string
      attName:
        type: string
      attUrl:
        type: string
      type:
        type: string
    type: object
...
info:
  contact:
    email: support@somewhere.com
    name: API Support
    url: http://somewhere.com/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://somewhere.com/
  title: CMS API
  version: "1.0"
paths:
...
  /api/v1/case:
    get:
      consumes:
      - application/json
      operationId: ListCase
      parameters:
      - default: 0
        description: start
        in: query
        name: start
        type: integer
      - default: 10
        description: length
        in: query
        name: length
        type: integer
      - description: start_date
        in: query
        name: start_date
        type: string
      - description: end_date
        in: query
        name: end_date
        type: string
      - description: caseType (can be comma-separated, e.g. 1,2,3)
        in: query
        name: caseType
        type: string
      - description: caseSType (can be comma-separated)
        in: query
        name: caseSType
        type: string
      - description: statusId (can be comma-separated)
        in: query
        name: statusId
        type: string
      - description: detail
        in: query
        name: detail
        type: string
      - description: caseId
        in: query
        name: caseId
        type: string
      - description: countryId (can be comma-separated)
        in: query
        name: countryId
        type: string
      - description: provId (can be comma-separated)
        in: query
        name: provId
        type: string
      - description: distId (can be comma-separated)
        in: query
        name: distId
        type: string
      - description: category (alias for statusId)
        in: query
        name: category
        type: string
      - description: createBy
        in: query
        name: createBy
        type: string
      - default: '"createdAt"'
        description: orderBy field name
        in: query
        name: orderBy
        type: string
      - default: '"DESC"'
        description: direction ASC or DESC
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: List Cases
      tags:
      - Cases
  /api/v1/case/{id}:
    delete:
      consumes:
      - application/json
      operationId: Delete Case
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Case
      tags:
      - Cases
    get:
      consumes:
      - application/json
      operationId: Case By Id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Cases By Id
      tags:
      - Cases
    patch:
      consumes:
      - application/json
      operationId: Update Case
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.CaseUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Case
      tags:
      - Cases
  /api/v1/case/add:
    post:
      consumes:
      - application/json
      operationId: Create Case
      parameters:
      - description: Create Data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.CaseInsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Create Case
      tags:
      - Cases
  /api/v1/case_history:
    get:
      consumes:
      - application/json
      operationId: Get Case History
      parameters:
      - default: 0
        description: start
        in: query
        name: start
        type: integer
      - default: 10
        description: length
        in: query
        name: length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Case History
      tags:
      - Cases
  /api/v1/case_history/{caseId}:
    get:
      consumes:
      - application/json
      operationId: Get Case History By Case Id
      parameters:
      - description: caseId
        in: path
        name: caseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Case History By Case Id
      tags:
      - Cases
  /api/v1/case_history/{id}:
    delete:
      consumes:
      - application/json
      operationId: Delete Case History
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Case History
      tags:
      - Cases
    patch:
      consumes:
      - application/json
      operationId: Update Case History
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.CaseHistoryUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Case History
      tags:
      - Cases
  /api/v1/case_history/add:
    post:
      consumes:
      - application/json
      operationId: Create Case History
      parameters:
      - description: Create Data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.CaseHistoryInsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Create Case History
      tags:
      - Cases
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"